Fix Actions
The `ctrl` actions were a thing I thought was working but it turns out
they were not.

After diving into the issue I found that it was in my custom sink, and
how it was mapping over the lines.
It turns out the first 'line' isnt actually a line its the action that
we want to take. And if we include this in the stripping we will lose
the ability to lost `ctrl` actions from fzf

We move the mapping to AFTER we take the action key, so that the user
specified action is preserved.